Fetch_AI/requirements.txt
langchain==0.1.0
langchain-core==0.1.10
langgraph==0.0.10
fetchai-sdk==0.16.3
flask==2.2.5
flask-cors==3.0.10
python-dotenv==1.0.0
tavily-python==0.2.6
qdrant-client==1.7.0
tiktoken==0.5.2